Make login and sessiontimeout views transient, redirect on login
By default, going to the login page creates a session and
an associated view, and when this session expires,
the view is deleted. So if you leave the login page long
enough to allow session expiration, then try to login,
this will fail because of a ViewExpiredException.

So make the view transient, which solves the problem.
Do the same to sessiontimeout.

On successful login, force a redirect, so that the
browser location shows welcome.xhtml.